Fix linting issues
